requirements.txt
pandas
psycopg2-binary
sqlalchemy
python-dotenv
streamlit
langchain
psycopg2
pytest